requirements.txt
# Analytics Assistant 项目依赖
# LangChain 生态使用最新 1.x 版本（2026-02 确认）

# ============================================
# 核心框架（LangChain 生态）
# ============================================
langchain>=1.2.0
langchain-core>=1.2.0
langchain-community>=0.4.1
langchain-openai>=1.1.0
langgraph>=1.0.9
langgraph-checkpoint>=2.0.0
langsmith>=0.4.0

# ============================================
# AI / ML
# ============================================
openai>=1.90.0
tiktoken>=0.9.0
faiss-cpu>=1.12.0
numpy>=2.0.0
chromadb>=1.0.15
jieba>=0.42.1
rank_bm25>=0.2.2

# ============================================
# Agent 框架（LangChain Deep Agents）
# ============================================
deepagents>=0.4.0

# ============================================
# Web 框架
# ============================================
fastapi>=0.115.0
uvicorn>=0.30.0
starlette>=0.40.0

# ============================================
# HTTP 客户端
# ============================================
httpx>=0.28.0
httpx-sse>=0.4.0
aiohttp>=3.10.0

# ============================================
# 数据验证 / 序列化
# ============================================
pydantic>=2.10.0
pydantic-settings>=2.5.0

# ============================================
# 安全 / 认证
# ============================================
cryptography>=45.0.0
PyJWT>=2.10.0

# ============================================
# 配置 / 工具
# ============================================
PyYAML>=6.0.0
certifi>=2024.0.0
typing_extensions>=4.10.0
python-dateutil>=2.9.0

# ============================================
# 测试依赖
# ============================================
hypothesis>=6.100.0
pytest>=8.0.0
